requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-dotenv==1.4.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
